Graphex

The Graph

data Graph a

Constructors

Graph

Fields

unGraph :: Map a (Set a)

attributes :: Map a (Map Text Text)

getInput :: FilePath -> IO (Graph Text)

Load a graph from a lookingglass JSON representation.

hGetInput :: Handle -> IO (Graph Text)

Working from an individual node in the graph.

directDepsOn :: Ord a => Graph a -> a -> Set a

Find the direct list of things that are referencing this module.

allDepsOn :: Ord a => Graph a -> a -> Set a

Flood fill to find all transitive dependencies on a starting module, excluding the original key.

why :: Ord a => Graph a -> a -> a -> Maybe (NonEmpty a)

Find an example path between two modules.

This is a short path, but the important part is that it represents how connectivity works.

Working on the graph as a whole.

reverseEdges :: Ord a => Graph a -> Graph a

Reverse all the arrows in the graphs.

rankings :: (NFData a, Ord a) => Graph a -> [(Int, a)]

Count the number of transitive dependencies for each module.

longest :: (NFData a, Ord a) => Graph a -> [a]

Find the longest path in a Graph.

allPathsTo :: Ord a => Graph a -> a -> a -> Graph a

Find all paths between two modules as a restricted graph of the intersection of reachable nodes from the start and to the end.

restrictTo :: Ord a => Graph a -> Set a -> Graph a

Restrict a graph to only the given set of modules.

mapMaybeWithKey :: (NFData a, NFData g) => (g -> Maybe a) -> Graph g -> [(g, a)]

Visit each node in the graph and apply a function to it.

filterNodes :: Ord a => (a -> Bool) -> Graph a -> Graph a

graphToDep :: Graph Text -> GraphDef

Convert a graph back to our dependency file format.

depToGraph :: GraphDef -> Graph Text

Convert a dependency file to a graph.

graphToTree :: Ord a => a -> Graph a -> Tree a

Convert a Graph to a Tree. If there is a cycle, treat the cycle point as a leaf.