Graphex.Core

Documentation

data Graph a

Constructors

Graph

Fields

unGraph :: Map a (Set a)

attributes :: Map a (Map Text Text)

singletonGraph :: Ord a => a -> a -> Graph a

mkGraph :: Ord a => a -> [a] -> Graph a

setAttribute :: Ord a => a -> Text -> Text -> Graph a -> Graph a

getAttribute :: Ord a => a -> Text -> Graph a -> Maybe Text

graphNodes :: Graph a -> [a]

data Import

Constructors

Import

Fields

module_ :: ModuleName

package :: Maybe Text

newtype ModuleName

Constructors

ModuleName

Fields

unModuleName :: Text

data ModulePath

Constructors

ModuleNoFile

ModuleFile FilePath

moduleFilePath :: Module -> Maybe FilePath

data Module

Constructors

Module

Fields

name :: ModuleName

path :: ModulePath

type ModuleGraph = Graph ModuleName

singletonModuleGraph :: ModuleName -> ModuleName -> ModuleGraph

mkModuleGraph :: ModuleName -> [ModuleName] -> ModuleGraph

convertGraph :: Ord b => (a -> b) -> Graph a -> Graph b

Convert a Graph of type a to a Graph of type b. This is like a functor, but it can't be a functor because of laws and stuff.