Graphex.Parser

Description

Implements a very crude Haskell import parser.

All it does is parse full module names that are imported in a file, ignoring false positives.

It does not use CPP, so all conditional imports are parsed out. For graphex, this makes sense in a way - those are all still dependencies.

Documentation

importParser :: MonadParsec e s m => Token s ~ Char => IsString (Tokens s) => m Import

importsParser :: forall e s m. MonadParsec e s m => Token s ~ Char => IsString (Tokens s) => m [Import]

removeBlockComments :: Text -> Text

blockCommentParser :: forall e s m. MonadParsec e s m => Token s ~ Char => IsString (Tokens s) => m String

parseFileImports :: FilePath -> IO [Import]