Graphex.Search

Documentation

bfsOn :: Ord r => (a -> r) -> (a -> [a]) -> a -> [a]

A BFS that will return a list of all reachable states.

The first argument is a representation function is used to deduplicate state. For some use cases where the entire state is valid, you can use id.

bfsWith :: Monoid s => (a -> s -> s) -> (a -> s -> Bool) -> (a -> [a]) -> a -> [a]

BFS with custom functions for remembering and recalling whether a state has been visited.

dfsOn :: Ord r => (a -> r) -> (a -> [a]) -> a -> [a]

A DFS variant of bfsOn.

dfsWith :: Monoid s => (a -> s -> s) -> (a -> s -> Bool) -> (a -> [a]) -> a -> [a]

A DFS variant of bfsWith.

findFirst :: Ord r => SearchFunction r a -> (a -> r) -> (a -> [a]) -> a -> (a -> Bool) -> Maybe a

Find the first match using the given search function (e.g., bfsOn or dfsOn).

flood :: Ord a => (a -> Set a) -> a -> Set a

Flood fill a graph from a starting point and return all visited points.