Graphex

Documentation

newtype Graph

Constructors

Graph

Fields

unGraph :: Map Text (Set Text)

reverseEdges :: Graph -> Graph

Reverse all the arrows in the graphs.

directDepsOn :: Graph -> Text -> Set Text

Find the direct list of things that are referencing this module.

allDepsOn :: Graph -> Text -> Set Text

Flood fill to find all transitive dependencies on a starting module, excluding the original key.

why :: Graph -> Text -> Text -> [Text]

Find an example path between two modules.

This is a short path, but the important part is that it represents how connectivy works.

rankings :: Graph -> Map Text Int

Count the number of transitive dependencies for each module.

allPathsTo :: Graph -> Text -> Text -> Graph

Find all paths between two modules as a restricted graph of the intersection of reachable nodes from the start and to the end.

restrictTo :: Graph -> Set Text -> Graph

Restrict a graph to only the given set of modules.

mapMaybeWithKey :: NFData a => (Text -> Maybe a) -> Graph -> [(Text, a)]

Visit each node in the graph and apply a function to it.

graphToDep :: Graph -> GraphDef

Convert a graph back to our dependency file format.

depToGraph :: GraphDef -> Graph

Convert a dependency file to a graph.

graphToTree :: Text -> Graph -> Tree Text

Convert a Graph to a Tree. If there is a cycle, treat the cycle point as a leaf.